// page/emun/coffee.enum.ts
export enum Coffee {
  ESPRESSO = 'Espresso',
  ESPRESSO_MOCCHIATO = 'Espresso Macchiatto',
  CAPPUCCINO = 'Cappuccino',
  MOCHA = 'Mocha',
  FLAT_WHITE = 'Flat White',
  AMERICANO = 'Americano',
  CAFE_LATTE = 'Cafe Latte',
  EXPRESSO_CON_PANNA = 'Espresso Con Panna',
  COFE_BREVE = 'Cofe Breve'
}